+++ lib/src/provider/signer.ts
@@ -1,4 +1,4 @@
-import { LedgerId, AccountId, SignerSignature, AccountBalance, AccountInfo, TransactionRecord, Transaction, Executable } from "@hashgraph/sdk";
+import { LedgerId, AccountId, SignerSignature, AccountBalance, AccountInfo, TransactionRecord, Transaction, Executable, Query, TransactionResponse, TransactionId } from "@hashgraph/sdk";
 import { Signer } from "@hashgraph/sdk/lib/Signer";
 import { HashConnect } from "../main";
 import { HashConnectProvider } from "./provider";
@@ -7,10 +7,14 @@ export class HashConnectSigner implements Signer {
 
     private hashconnect: HashConnect;
     private provider: HashConnectProvider;
+    private topicId: string;
+    private accountToSign: string | AccountId;
 
-    constructor(hashconnect: HashConnect, provider: HashConnectProvider) {
+    constructor(hashconnect: HashConnect, provider: HashConnectProvider, accountToSign: string, topic: string) {
         this.hashconnect = hashconnect;
         this.provider = provider;
+        this.accountToSign = accountToSign;
+        this.topicId = topic;
     }
 
     getLedgerId: () => LedgerId | null;
@@ -21,12 +25,47 @@ export class HashConnectSigner implements Signer {
     getAccountBalance: () => Promise<AccountBalance>;
     getAccountInfo: () => Promise<AccountInfo>;
     getAccountRecords: () => Promise<TransactionRecord[]>;
-    signTransaction: (transaction: Transaction) => Promise<Transaction>;
-    checkTransaction: (transaction: Transaction) => Promise<Transaction>;
-    populateTransaction: (transaction: Transaction) => Promise<Transaction>;
+    async signTransaction(transaction: Transaction): Promise<Transaction> {
+        return transaction.freezeWith(this.provider.client)
+    };
 
-    sendRequest<RequestT, ResponseT, OutputT>(request: Executable<RequestT, ResponseT, OutputT>): Promise<OutputT> {
-        return this.provider.sendRequest(request);
+    checkTransaction: (transaction: Transaction) => Promise<Transaction>;
+    
+    async populateTransaction(transaction: Transaction): Promise<Transaction> {
+        // await this.checkTransaction(transaction);
+
+        transaction.setTransactionId(TransactionId.generate(this.accountToSign));
+        // transaction.setNodeAccountIds([]);
+
+        return transaction;
+    };
+
+    async sendRequest<RequestT, ResponseT, OutputT>(request: Executable<RequestT, ResponseT, OutputT>): Promise<OutputT> {
+        
+        const transaction = {
+            byteArray: this.getBytesOf(request),
+            metadata: {
+                accountToSign: this.accountToSign.toString(),
+                returnTransaction: false,
+            },
+            topic: this.topicId,
+        };
+
+        let res = await this.hashconnect.sendTransaction(this.topicId, transaction);
+        
+        let response: TransactionResponse = res.response as TransactionResponse;
+        
+        return (response as unknown) as OutputT;
     }
+
+    private getBytesOf<RequestT, ResponseT, OutputT>(request: Executable<RequestT, ResponseT, OutputT>): Uint8Array {
+        console.log(request);
+        console.log(typeof(request))
+        return ((request as unknown) as Transaction).toBytes();
+        if (request instanceof Transaction || request instanceof Query) {
+        } else {
+          throw new Error("Only Transactions and Queries can be serialized to be sent for signing by the HashPack wallet.");
+        }
+      }
 
 }