+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.2.0
+- Added provider/signer (HIP-338) functionality
+- Added ability to also get transaction record
+- Added transaction response in addition to receipt
+  
 ## v0.1.5
 - Refactored authentication to be more secure